Implement BaseMaskedArray class for integer/boolean ExtensionArrays
